+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.6.2] - 2019-05-22
 ### Fixed
 - Calculated price with taxes in order
+- Calculated price with taxes in quote
 
 ## [1.6.1] - 2019-05-21
 ### Fixed